Code refactor
Renaming models. Removing id repeated property from all models.
