+++ src/models/location.model.js
@@ -52,7 +52,11 @@ const locationSchema = new Schema({
 
 locationSchema.set('toJSON', {
   getters: true,
-  virtuals: true
+  virtuals: true,
+  transform: (doc, ret) => {
+    delete ret.id
+    return ret
+  }
 })
 
 module.exports = mongoose.model(mongoDB.Model.Location, locationSchema)
