+++ src/models/schedule-types.enum.js
