+++ src/models/user.model.js
@@ -2,7 +2,7 @@ const mongoose = require('mongoose')
 const Schema = mongoose.Schema
 const bcrypt = require('bcrypt-nodejs')
 const mongoDB = require('../config/mongoose.collections.json')
-const roles = require('./roles.server.enum')()
+const roles = require('./roles.enum.js')()
 const config = require('../config/config')()
 const Promise = require('bluebird')
 
@@ -142,6 +142,7 @@ userSchema.set('toJSON', {
   virtuals: true,
   transform: (doc, ret) => {
     delete ret.password
+    delete ret.id
     return ret
   }
 })
